[LLCL] Change WorkQueue::getParallelismLevel to return size_t (#2331)
The return type being `int` does not make sense from a type or semantic
perspective - the pool size cannot be negative. Change the return type to be
`size_t` and adjust callers appropriately to remove the casts downstream.

Fixes #2324.

MODULAR_ORIG_COMMIT_REV_ID: bb181524bfc0c3dbbb97e7ab7822f707fdecee0e

+++ LLCL/include/LLCL/Runtime/WorkQueue.h
@@ -48,7 +48,7 @@ public:
   /// Return the pool size maintained by this work queue. Kernels can use
   /// this as a hint indicating the maximum useful number of work items
   /// they should break themselves into.
-  virtual int getParallelismLevel() const = 0;
+  virtual size_t getParallelismLevel() const = 0;
 
   /// Shutdown the thread pool and quiesce in preparation for destruction.
   virtual void shutdown() = 0;
